Updated ImageCarousel.jsx
Created Slide.js with mocked data and then imported it in ImageCarousel.jsx
